requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
gTTS==2.5.4
idna==3.10
macholib==1.16.3
packaging==24.2
pygame==2.6.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
soupsieve==2.6
spotipy==2.24.0
urllib3==2.3.0
